Move mostRecentTuple
