+++ libs/wire-api-federation/src/Wire/API/Federation/Version.hs
@@ -41,6 +41,7 @@ where
 
 import Control.Lens (makeLenses, (?~))
 import Data.Aeson (FromJSON (..), ToJSON (..))
+import Data.List.NonEmpty qualified as NE
 import Data.OpenApi qualified as S
 import Data.Schema
 import Data.Set qualified as Set
@@ -140,6 +141,21 @@ enumVersionRange =
 latestCommonVersion :: VersionRange -> Set Int -> Maybe Version
 latestCommonVersion (Set.map versionInt . enumVersionRange -> localVersions) remoteVersions =
   intToVersion =<< Set.lookupMax (Set.intersection localVersions remoteVersions)
+
+mostRecentTuple :: forall a. (a -> Maybe VersionRange) -> NE.NonEmpty a -> Set Int -> Maybe (a, Version)
+mostRecentTuple pr (NE.toList -> as) remoteVersions = foldl' combine Nothing as
+  where
+    combine :: Maybe (a, Version) -> a -> Maybe (a, Version)
+    combine greatest a =
+      let notifGreatest = pr a >>= flip latestCommonVersion remoteVersions
+       in case (greatest, notifGreatest) of
+            (Nothing, Nothing) -> Nothing
+            (Nothing, Just v) -> Just (a, v)
+            (Just (gn, gv), Nothing) -> Just (gn, gv)
+            (Just (gn, gv), Just v) ->
+              if v > gv
+                then Just (a, v)
+                else Just (gn, gv)
 
 $(genSingletons [''Version])
